antes das férias

+++ Code/app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -7,17 +7,25 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca) {
+        $this->marca = $marca;
+    }
     /**
      * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
      */
     public function index()
     {
-        $marcas = Marca::all();
-        return $marcas;
+        //$marcas = Marca::all();
+        $marcas = $this->marca->all();
+        return response()->json($marcas, 200);
     }
 
     /**
      * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
      */
     public function create()
     {
@@ -26,24 +34,38 @@ class MarcaController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
-
-        return $marca;
+        //$marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
+        return response()->json($marca, 201);
     }
 
     /**
      * Display the specified resource.
+     *
+     * @param  Integer
+     * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+        if($marca === null) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe'], 404) ;
+        }
+
+        return response()->json($marca, 200);
     }
 
     /**
      * Show the form for editing the specified resource.
+     *
+     * @param  \App\Models\Marca  $marca
+     * @return \Illuminate\Http\Response
      */
     public function edit(Marca $marca)
     {
@@ -52,23 +74,39 @@ class MarcaController extends Controller
 
     /**
      * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  Integer
+     * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
-       // print_r($request->all());
-       // echo '<hr>';
-       // print_r($marca->getAttributes());
-       $marca->update($request->all());
-       return $marca;
+        $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'], 404);
+        }
+
+        $marca->update($request->all());
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
+     *
+     * @param  Integer
+     * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if($marca === null) {
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'], 404);
+        }
 
         $marca->delete();
-        return ['A marca foi removida com sucesso'];
+        return response()->json(['msg' => 'A marca foi removida com sucesso!'], 200);
+
     }
 }
